journalpump: refactored senders to separate modules
