+++ journalpump/journalpump.py
@@ -4,62 +4,27 @@
 # See the file `LICENSE` for details.
 
 from . import geohash, statsd
-from . daemon import ServiceDaemon
-from . rsyslog import SyslogTcpClient
-from . util import atomic_replace_file, get_requests_session
+from .daemon import ServiceDaemon
+from .util import atomic_replace_file, default_json_serialization
 from functools import reduce
-from io import BytesIO
-from kafka import KafkaProducer
 from systemd.journal import Reader
-from threading import Lock, Thread
+from .senders.base import MAX_KAFKA_MESSAGE_SIZE
+from .senders.base import Tagged
+from .senders import AWSCloudWatchSender, RsyslogSender, LogplexSender, FileSender, KafkaSender, ElasticsearchSender
 import copy
 import datetime
 import json
-import kafka.errors
 import logging
-import random
 import re
-import requests
 import select
-import socket
 import systemd.journal
 import time
 import uuid
-import botocore
-import boto3
-
-try:
-    import snappy
-except ImportError:
-    snappy = None
 
 try:
     from geoip2.database import Reader as GeoIPReader
 except ImportError:
     GeoIPReader = None
-
-
-KAFKA_CONN_ERRORS = tuple(kafka.errors.RETRY_ERROR_TYPES) + (
-    kafka.errors.UnknownError,
-    socket.timeout,
-)
-
-RSYSLOG_CONN_ERRORS = (socket.timeout, ConnectionRefusedError)
-
-KAFKA_COMPRESSED_MESSAGE_OVERHEAD = 30
-MAX_KAFKA_MESSAGE_SIZE = 1024 ** 2  # 1 MiB
-
-MAX_ERROR_MESSAGES = 8
-MAX_ERROR_MESSAGE_LEN = 128
-
-logging.getLogger("kafka").setLevel(logging.CRITICAL)  # remove client-internal tracebacks from logging output
-
-
-# Map running/_connected to health
-def _convert_to_health(*, running, connected):
-    if running:
-        return ("connected", 3) if connected else ("disconnected", 2)
-    return ("stale", 2) if connected else("stopped", 0)
 
 
 def _convert_uuid(s):
@@ -72,13 +37,6 @@ def convert_mon(s):  # pylint: disable=unused-argument
 
 def convert_realtime(t):
     return int(t) / 1000000.0  # Stock systemd transforms these into datetimes
-
-
-def default_json_serialization(obj):  # pylint: disable=inconsistent-return-statements
-    if isinstance(obj, bytes):
-        return obj.decode("utf8")
-    if isinstance(obj, datetime.datetime):
-        return obj.isoformat()
 
 
 converters = {
@@ -138,662 +96,6 @@ class PumpReader(Reader):
                 return JournalObject(cursor=self._get_cursor(), entry=self.convert_entry(entry))
 
         return None
-
-
-class Tagged:
-    def __init__(self, tags=None, **kw):
-        self._tags = (tags or {}).copy()
-        self._tags.update(kw)
-
-    def make_tags(self, tags=None):
-        output = self._tags.copy()
-        output.update(tags or {})
-        return output
-
-    def replace_tags(self, tags):
-        self._tags = tags
-
-
-class LogSender(Thread, Tagged):
-    def __init__(self, *, name, reader, config, field_filter, stats, max_send_interval,
-                 extra_field_values=None, tags=None, msg_buffer_max_length=50000):
-        Thread.__init__(self)
-        Tagged.__init__(self, tags, sender=name)
-        self.log = logging.getLogger("LogSender:{}".format(reader.name))
-        self.name = name
-        self.stats = stats
-        self.config = config
-        self.extra_field_values = extra_field_values
-        self.field_filter = field_filter
-        self.msg_buffer_max_length = msg_buffer_max_length
-        self.last_maintenance_fail = 0
-        self.last_send_time = time.monotonic()
-        self.max_send_interval = max_send_interval
-        self.running = True
-        self._sent_cursor = None
-        self._sent_count = 0
-        self._sent_bytes = 0
-        self._connected = False
-        self._connected_changed = time.monotonic()  # last time _connected status changed
-        self._errors = []
-        self.msg_buffer = MsgBuffer()
-        self._backoff_attempt = 0
-        self.log.info("Initialized %s", self.__class__.__name__)
-
-    def _backoff(self, *, base=0.5, cap=1800.0):
-        self._backoff_attempt += 1
-        t = min(cap, base * 2 ** self._backoff_attempt) / 2
-        t = random.random() * t + t
-        self.log.info("Sleeping for %.0f seconds", t)
-        time.sleep(t)
-
-    def refresh_stats(self):
-        tags = self.make_tags()
-        self.stats.gauge("journal.last_sent_ago", value=time.monotonic() - self.last_send_time, tags=tags)
-        self.stats.gauge("journal.sent_bytes", value=self._sent_bytes, tags=tags)
-        self.stats.gauge("journal.sent_lines", value=self._sent_count, tags=tags)
-        self.stats.gauge("journal.status",
-                         value=_convert_to_health(running=self.running, connected=self._connected)[1], tags=tags)
-
-    def get_state(self):
-        return {
-            "type": self.__class__.__name__,
-            "buffer": {
-                "cursor": self.msg_buffer.cursor,
-                "total_size": self.msg_buffer.total_size,
-                "entry_num": self.msg_buffer.entry_num,
-                "current_queue": len(self.msg_buffer.messages),
-            },
-            "sent": {
-                "cursor": self._sent_cursor,
-                "count": self._sent_count,
-                "bytes": self._sent_bytes,
-            },
-            "health": {
-                "status": _convert_to_health(running=self.running, connected=self._connected)[0],
-                "elapsed": time.monotonic() - self._connected_changed,
-                "errors": self._errors,
-            }
-        }
-
-    def mark_connected(self):
-        if not self._connected:
-            self._connected_changed = time.monotonic()
-        self._connected = True
-        self._errors = []
-        self._backoff_attempt = 0
-
-    def mark_disconnected(self, error=None):
-        if self._connected:
-            self._connected_changed = time.monotonic()
-        self._connected = False
-
-        if error is None:
-            return
-
-        if isinstance(error, str):
-            msg = error
-        elif isinstance(error, Exception):
-            msg = getattr(error, 'message', repr(error))
-        else:
-            msg = repr(error)
-
-        if not msg:
-            return
-
-        msg = msg[:MAX_ERROR_MESSAGE_LEN]
-        if msg in self._errors:
-            return
-        self._errors.append(msg)
-
-        if len(self._errors) > MAX_ERROR_MESSAGES:
-            self._errors.pop(1)  # always keep the first error message
-
-    def mark_sent(self, *, messages, cursor):
-        self._sent_count += len(messages)
-        self._sent_bytes += sum(len(m) for m in messages)
-        self._sent_cursor = cursor
-        self.mark_connected()
-
-    def request_stop(self):
-        self.running = False
-
-    def send_messages(self, *, messages, cursor):
-        pass
-
-    def maintenance_operations(self):
-        # This can be overridden in the classes that inherit this
-        pass
-
-    def run(self):
-        while self.running:
-            try:
-                # Don't run maintenance operations again immediately if it just failed
-                if not self.last_maintenance_fail or time.monotonic() - self.last_maintenance_fail > 60:
-                    self.maintenance_operations()
-            except Exception as ex:  # pylint: disable=broad-except
-                self.log.error("Maintenance operation failed: %r", ex)
-                self.stats.unexpected_exception(ex=ex, where="maintenance_operation")
-                self.last_maintenance_fail = time.monotonic()
-            if len(self.msg_buffer) > 1000 or \
-               time.monotonic() - self.last_send_time > self.max_send_interval:
-                self.get_and_send_messages()
-            else:
-                time.sleep(0.1)
-        self.log.info("Stopping")
-
-    def get_and_send_messages(self):
-        start_time = time.monotonic()
-        msg_count = None
-        try:
-            messages = self.msg_buffer.get_items()
-            msg_count = len(messages)
-            self.log.debug("Got %d items from msg_buffer", msg_count)
-
-            while self.running and messages:
-                batch_size = len(messages[0][0]) + KAFKA_COMPRESSED_MESSAGE_OVERHEAD
-                index = 1
-                while index < len(messages):
-                    item_size = len(messages[index][0]) + KAFKA_COMPRESSED_MESSAGE_OVERHEAD
-                    if batch_size + item_size >= MAX_KAFKA_MESSAGE_SIZE:
-                        break
-                    batch_size += item_size
-                    index += 1
-
-                messages_batch = messages[:index]
-                message_bodies = [m[0] for m in messages_batch]
-                if self.send_messages(messages=message_bodies, cursor=messages_batch[-1][1]):
-                    messages = messages[index:]
-
-            self.log.debug("Sending %d msgs, took %.4fs", msg_count, time.monotonic() - start_time)
-            self.last_send_time = time.monotonic()
-        except Exception:   # pylint: disable=broad-except
-            self.log.exception("Problem sending %r messages", msg_count)
-            self._backoff()
-
-
-class AWSCloudWatchSender(LogSender):
-    def __init__(self, *, config, aws_cloudwatch_logs=None, **kwargs):
-        super().__init__(
-            config=config,
-            max_send_interval=config.get("max_send_interval", 0.3),
-            **kwargs
-        )
-        self._logs = aws_cloudwatch_logs
-        self.log_group = self.config.get("aws_cloudwatch_log_group")
-        self.log_stream = self.config.get("aws_cloudwatch_log_stream")
-        self._next_sequence_token = None
-        self._init_logs()
-
-    def _init_logs(self):
-        if self._logs is None:
-            if self.log_group is None or self.log_stream is None:
-                raise Exception("AWS CloudWatch log group and stream names need to be configured")
-            kwargs = {}
-            if self.config.get("aws_region") is not None:
-                kwargs["region_name"] = self.config.get("aws_region")
-            if self.config.get("aws_access_key_id") is not None:
-                kwargs["aws_access_key_id"] = self.config.get("aws_access_key_id")
-            if self.config.get("aws_secret_access_key") is not None:
-                kwargs["aws_secret_access_key"] = self.config.get("aws_secret_access_key")
-            self._logs = boto3.client("logs", **kwargs)
-        # Create the log group and stream if they don't exist yet
-        try:
-            self._logs.create_log_group(logGroupName=self.log_group)
-        except botocore.exceptions.ClientError as err:
-            if err.response["Error"]["Code"] != "ResourceAlreadyExistsException":
-                raise
-        try:
-            self._logs.create_log_stream(logGroupName=self.log_group, logStreamName=self.log_stream)
-        except botocore.exceptions.ClientError as err:
-            if err.response["Error"]["Code"] != "ResourceAlreadyExistsException":
-                raise
-        streams = self._logs.describe_log_streams(logGroupName=self.log_group).get("logStreams")
-        if streams is not None:
-            stream_metadata = [stream for stream in streams if stream['logStreamName'] == self.log_stream][0]
-            self._next_sequence_token = stream_metadata.get("uploadSequenceToken")
-            self.mark_connected()
-        else:
-            raise Exception("AWS CloudWatch logs could not update sequence token")
-
-    def send_messages(self, *, messages, cursor):
-        log_events = []
-        for msg in messages:
-            raw_message = msg.decode("utf8")
-            message = json.loads(raw_message)
-            timestamp = message.get("REALTIME_TIMESTAMP") or time.time()
-            log_events.append(
-                {
-                    "timestamp": int(timestamp * 1000.0),
-                    "message": raw_message
-                }
-            )
-        kwargs = {
-            "logGroupName": self.log_group,
-            "logStreamName": self.log_stream,
-            "logEvents": log_events
-        }
-        if self._next_sequence_token is not None:
-            kwargs["sequenceToken"] = self._next_sequence_token
-        try:
-            response = self._logs.put_log_events(**kwargs)
-        except botocore.exceptions.ClientError as err:
-            err_code = err.response["Error"]["Code"]
-            err_msg = err.response["Error"]["Message"]
-            self.mark_disconnected()
-            self.log.error("Error sending events %r: %r", err_code, err_msg)
-            self.stats.unexpected_exception(ex=err, where="sender", tags=self.make_tags({"app": "journalpump"}))
-            self._backoff()
-            self._init_logs()
-        except Exception as ex:  # pylint: disable=broad-except
-            self.mark_disconnected(ex)
-            self.log.exception("Unexpected exception during send to AWS CloudWatch")
-            self.stats.unexpected_exception(ex=ex, where="sender", tags=self.make_tags({"app": "journalpump"}))
-            self._backoff()
-            self._init_logs()
-        else:
-            if 200 <= response["ResponseMetadata"]["HTTPStatusCode"] < 300:
-                self.mark_sent(messages=messages, cursor=cursor)
-                self._next_sequence_token = response["nextSequenceToken"]
-                return True
-        return False
-
-
-class KafkaSender(LogSender):
-    def __init__(self, *, config, **kwargs):
-        super().__init__(config=config, max_send_interval=config.get("max_send_interval", 0.3), **kwargs)
-        self.kafka_producer = None
-        self.kafka_msg_key = self.config.get("kafka_msg_key")
-        if self.kafka_msg_key:
-            self.kafka_msg_key = self.kafka_msg_key.encode("utf8")
-        self.topic = self.config.get("kafka_topic")
-
-    def _init_kafka(self):
-        self.log.info("Initializing Kafka client, address: %r", self.config["kafka_address"])
-        self.mark_disconnected()
-        while self.running:
-            try:
-                if self.kafka_producer:
-                    self.kafka_producer.close()
-
-                self.kafka_producer = KafkaProducer(
-                    api_version=self.config.get("kafka_api_version", "0.9"),
-                    bootstrap_servers=self.config.get("kafka_address"),
-                    compression_type="snappy" if snappy else "gzip",
-                    linger_ms=500,  # wait up 500 ms to see if we can send msgs in a group
-                    reconnect_backoff_ms=1000,  # up from the default 50ms to reduce connection attempts
-                    reconnect_backoff_max_ms=10000,  # up the upper bound for backoff to 10 seconds
-                    security_protocol="SSL" if self.config.get("ssl") is True else "PLAINTEXT",
-                    ssl_cafile=self.config.get("ca"),
-                    ssl_certfile=self.config.get("certfile"),
-                    ssl_keyfile=self.config.get("keyfile"),
-                )
-                self.log.info("Initialized Kafka Client, address: %r", self.config["kafka_address"])
-                self.mark_connected()
-                break
-            except KAFKA_CONN_ERRORS as ex:
-                self.mark_disconnected(ex)
-                self.log.warning("Retriable error during Kafka initialization: %s: %s", ex.__class__.__name__, ex)
-                self._backoff()
-            self.kafka_producer = None
-            time.sleep(5.0)
-
-    def send_messages(self, *, messages, cursor):
-        if not self.kafka_producer:
-            self._init_kafka()
-        try:
-            for msg in messages:
-                self.kafka_producer.send(topic=self.topic, value=msg, key=self.kafka_msg_key)
-            self.kafka_producer.flush()
-            self.mark_sent(messages=messages, cursor=cursor)
-            return True
-        except KAFKA_CONN_ERRORS as ex:
-            self.mark_disconnected(ex)
-            self.log.info("Kafka retriable error during send: %s: %s, waiting", ex.__class__.__name__, ex)
-            self._backoff()
-            self._init_kafka()
-        except Exception as ex:  # pylint: disable=broad-except
-            self.mark_disconnected(ex)
-            self.log.exception("Unexpected exception during send to kafka")
-            self.stats.unexpected_exception(ex=ex, where="sender", tags=self.make_tags({"app": "journalpump"}))
-            self._backoff()
-            self._init_kafka()
-        return False
-
-
-class FileSender(LogSender):
-    def __init__(self, *, config, **kwargs):
-        self.mark_disconnected()
-        super().__init__(config=config, max_send_interval=config.get("max_send_interval", 0.3), **kwargs)
-        self.output = open(config["file_output"], "ab")
-        self.mark_connected()
-
-    def send_messages(self, *, messages, cursor):
-        for msg in messages:
-            self.output.write(msg + b"\n")
-
-        self.mark_sent(messages=messages, cursor=cursor)
-        return True
-
-
-class RsyslogSender(LogSender):
-    def __init__(self, *, config, **kwargs):
-        super().__init__(config=config, max_send_interval=config.get("max_send_interval", 0.3), **kwargs)
-        self.rsyslog_client = None
-        self.sd = None
-        self.default_facility = 1
-        self.default_severity = 6
-
-    def _init_rsyslog_client(self):
-        self.log.info("Initializing Rsyslog Client")
-        self.mark_disconnected()
-        while self.running:
-            try:
-                if self.rsyslog_client:
-                    self.rsyslog_client.close()
-
-                self.default_facility = self.config.get("default_facility", 1)
-                self.default_severity = self.config.get("default_severity", 6)
-                self.sd = self.config.get("structured_data")
-
-                server = self.config.get("rsyslog_server")
-                port = self.config.get("rsyslog_port", 514)
-
-                self.rsyslog_client = SyslogTcpClient(
-                    server=server,
-                    port=port,
-                    rfc=self.config.get("format", "rfc5424").upper(),
-                    protocol="SSL" if self.config.get("ssl") is True else "PLAINTEXT",
-                    cacerts=self.config.get("ca_certs"),
-                    keyfile=self.config.get("client_key"),
-                    certfile=self.config.get("client_cert"),
-                    log_format=self.config.get("logline"),
-                )
-                self.log.info("Initialized Rsyslog Client, server: %s, port: %d", server, port)
-                self.mark_connected()
-                break
-            except RSYSLOG_CONN_ERRORS as ex:
-                self.log.warning(
-                    "Retryable error during Rsyslog Client initialization: %s: %s, sleeping", ex.__class__.__name__, ex
-                )
-                self.mark_disconnected(ex)
-                self._backoff()
-            self.rsyslog_client = None
-            time.sleep(5.0)
-
-    def send_messages(self, *, messages, cursor):
-        if not self.rsyslog_client:
-            self._init_rsyslog_client()
-        try:
-            for msg in messages:
-                message = json.loads(msg.decode("utf8"))
-                _facility = message.get("SYSLOG_FACILITY")
-                try:
-                    facility = int(_facility[0] if isinstance(_facility, list) else _facility)
-                except Exception:  # pylint: disable=broad-except
-                    facility = self.default_facility
-
-                severity = int(message.get("PRIORITY", self.default_severity))
-                timestamp = message["timestamp"][:26] + "Z"  # Assume UTC for now
-                hostname = message.get("HOSTNAME")
-                appname = message.get("SYSLOG_IDENTIFIER", message.get("SYSTEMD_UNIT", message.get('PROCESS_NAME')))
-                progid = message.get("PID")
-                txt = message.get("MESSAGE")
-
-                self.rsyslog_client.log(
-                    facility=facility,
-                    severity=severity,
-                    timestamp=timestamp,
-                    hostname=hostname,
-                    program=appname,
-                    pid=progid,
-                    msg=txt,
-                    sd=self.sd
-                )
-            self.mark_sent(messages=messages, cursor=cursor)
-            return True
-        except RSYSLOG_CONN_ERRORS as ex:
-            self.mark_disconnected(ex)
-            self.log.info("Rsyslog Client retryable error during send: %s: %s, waiting", ex.__class__.__name__, ex)
-            self._backoff()
-            self._init_rsyslog_client()
-        except Exception as ex:  # pylint: disable=broad-except
-            self.mark_disconnected(ex)
-            self.log.exception("Unexpected exception during send to rsyslog")
-            self.stats.unexpected_exception(ex=ex, where="sender", tags=self.make_tags({"app": "journalpump"}))
-            self._backoff()
-            self._init_rsyslog_client()
-        return False
-
-
-class ElasticsearchSender(LogSender):
-    def __init__(self, *, config, **kwargs):
-        super().__init__(config=config, max_send_interval=config.get("max_send_interval", 10.0), **kwargs)
-        self.session_url = self.config.get("elasticsearch_url")
-        self.last_index_check_time = 0
-        self.request_timeout = self.config.get("elasticsearch_timeout", 10.0)
-        self.index_days_max = self.config.get("elasticsearch_index_days_max", 3)
-        self.index_name = self.config.get("elasticsearch_index_prefix", "journalpump")
-        self.session = get_requests_session()
-        # If ca is set in config we use that, otherwise we verify using builtin CA cert list
-        self.session.verify = self.config.get("ca", True)
-        self.indices = set()
-        self.last_es_error = None
-
-    def _init_es_client(self):
-        if self.indices:
-            return True
-        self.mark_disconnected()
-        try:
-            self.indices = set(self.session.get(self.session_url + "/_aliases", timeout=60.0).json().keys())
-            self.last_es_error = None
-        except (requests.exceptions.ConnectionError, requests.exceptions.Timeout) as ex:
-            self.mark_disconnected(ex)
-            if ex.__class__ != self.last_es_error.__class__:
-                # only log these errors once, not every 10 seconds
-                self.log.warning("ES connection error: %s: %s", ex.__class__.__name__, ex)
-
-            self.last_es_error = ex
-            return False
-        except Exception as ex:  # pylint: disable=broad-except
-            self.mark_disconnected(ex)
-            self.log.exception("Unexpected exception connecting to ES")
-            self.stats.unexpected_exception(ex, where="es_pump_init_es_client")
-            return False
-
-        self.log.info("Initialized ES connection")
-        self.mark_connected()
-        return True
-
-    def create_index_and_mappings(self, index_name):
-        try:
-            self.log.info("Creating index: %r", index_name)
-            res = self.session.put(
-                self.session_url + "/{}?include_type_name=true".format(index_name),
-                json={
-                    "mappings": {
-                        "journal_msg": {
-                            "properties": {
-                                "SYSTEMD_SESSION": {"type": "text"},
-                                "SESSION_ID": {"type": "text"},
-                            }
-                        }
-                    },
-                },
-                timeout=60.0,
-            )
-            if res.status_code in (200, 201) or "already_exists_exception" in res.text:
-                self.indices.add(index_name)
-            else:
-                self.mark_disconnected('Cannot create index "{index_name}" ({res.status_code} {res.text})')
-                self.log.warning("Could not create index mappings for: %r, %r %r", index_name, res.text, res.status_code)
-        except (requests.exceptions.ConnectionError, requests.exceptions.Timeout) as ex:
-            self.mark_disconnected(ex)
-            self.log.error("Problem creating index %r: %s: %s", index_name, ex.__class__.__name__, ex)
-            self.stats.unexpected_exception(ex, where="es_pump_create_index_and_mappings")
-            self._backoff()
-
-    @staticmethod
-    def format_message_for_es(buf, header, message):
-        buf.write(json.dumps(header, default=default_json_serialization).encode("utf8") + b"\n")
-        # Message already in utf8 encoded bytestring form
-        buf.write(message + b"\n")
-
-    def check_indices(self):
-        aliases = self.session.get(self.session_url + "/_aliases", timeout=60.0).json()
-        index_full_prefix = "{}-".format(self.index_name)
-        indices = sorted(key for key in aliases.keys() if key.startswith(index_full_prefix))
-        self.log.info("Checking indices, currently: %r are available, max_indices: %r", indices, self.index_days_max)
-        while len(indices) > self.index_days_max:
-            index_to_delete = indices.pop(0)
-            self.log.info("Deleting index: %r since we only keep %d days worth of indices",
-                          index_to_delete, self.index_days_max)
-            try:
-                self.session.delete("{}/{}".format(self.session_url, index_to_delete), timeout=60.0)
-                self.indices.discard(index_to_delete)
-            except Exception as ex:   # pylint: disable=broad-except
-                self.log.exception("Unexpected exception deleting index %r", index_to_delete)
-                self.stats.unexpected_exception(ex, where="es_pump_check_indices")
-
-    def maintenance_operations(self):
-        if time.monotonic() - self.last_index_check_time > 3600:
-            self.last_index_check_time = time.monotonic()
-            self.check_indices()
-
-    def send_messages(self, *, messages, cursor):
-        buf = BytesIO()
-        start_time = time.monotonic()
-        try:
-            es_available = self._init_es_client()
-            if not es_available:
-                self.log.warning("Waiting for ES connection")
-                self._backoff()
-                return False
-            for msg in messages:
-                message = json.loads(msg.decode("utf8"))
-                # ISO datetime's first 10 characters are equivalent to the date we need i.e. '2018-04-14'
-                index_name = "{}-{}".format(self.index_name, message["timestamp"][:10])
-                if index_name not in self.indices:
-                    self.create_index_and_mappings(index_name)
-
-                header = {
-                    "index": {
-                        "_index": index_name,
-                        "_type": "journal_msg",
-                    }
-                }
-                self.format_message_for_es(buf, header, msg)
-
-            # If we have messages, send them along to ES
-            if buf.tell():
-                buf_size = buf.tell()
-                buf.seek(0)
-                # Elasticsearch allows using _index even when posting
-                # to particular index to override the index the entry
-                # so this is mostly cosmetic to avoid needing to
-                # expose /_bulk. We use simply the most recent index
-                # name as base as it does not really matter which one
-                # we use.
-                res = self.session.post(
-                    '%s/%s/_bulk' % (self.session_url, index_name),
-                    data=buf,
-                    headers={
-                        "content-length": str(buf_size),
-                        "content-type": "application/x-ndjson",
-                    },
-                    timeout=120.0,
-                )
-                buf.seek(0)
-                buf.truncate(0)
-
-                self.mark_sent(messages=messages, cursor=cursor)
-                self.log.info("Sent %d log events to ES successfully: %r, took: %.2fs",
-                              len(messages), res.status_code in {200, 201}, time.monotonic() - start_time)
-        except Exception as ex:  # pylint: disable=broad-except
-            self.mark_disconnected(ex)
-            short_msg = str(ex)[:200]
-            self.log.exception("Problem sending logs to ES: %s: %s", ex.__class__.__name__, short_msg)
-            self._backoff()
-            return False
-
-        return True
-
-
-class LogplexSender(LogSender):
-    def __init__(self, *, config, **kwargs):
-        super().__init__(config=config, max_send_interval=config.get("max_send_interval", 5.0), **kwargs)
-        self.logplex_input_url = config["logplex_log_input_url"]
-        self.request_timeout = config.get("logplex_request_timeout", 2)
-        self.logplex_token = config["logplex_token"]
-        self.session = get_requests_session()
-        self.msg_id = "-"
-        self.structured_data = "-"
-        self.mark_connected()
-
-    def format_msg(self, msg):
-        # TODO: figure out a way to optionally get the entry without JSON
-        entry = json.loads(msg.decode("utf8"))
-        hostname = entry.get("_HOSTNAME", "localhost")
-        pid = entry.get("_PID", "localhost")
-        pkt = "<190>1 {} {} {} {} {} {}".format(
-            datetime.datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%S+00:00 "),
-            hostname,
-            self.logplex_token,
-            pid,
-            self.msg_id,
-            self.structured_data)
-        pkt += entry["MESSAGE"]
-        pkt = pkt.encode("utf8")
-        return '{} {}'.format(len(pkt), pkt)
-
-    def send_messages(self, *, messages, cursor):
-        auth = ('token', self.logplex_token)
-        msg_data = ''.join([self.format_msg(msg) for msg in messages])
-        msg_count = len(messages)
-        headers = {
-            "Content-Type": "application/logplex-1",
-            "Logplex-Msg-Count": msg_count,
-        }
-        self.session.post(
-            self.logplex_input_url,
-            auth=auth,
-            headers=headers,
-            data=msg_data,
-            timeout=self.request_timeout,
-            verify=False
-        )
-        self.mark_sent(messages=messages, cursor=cursor)
-
-
-class MsgBuffer:
-    def __init__(self):
-        self.log = logging.getLogger("MsgBuffer")
-        self.messages = []
-        self.lock = Lock()
-        self.entry_num = 0
-        self.total_size = 0
-        self.last_journal_msg_time = time.monotonic()
-        self.cursor = None
-
-    def __len__(self):
-        return len(self.messages)
-
-    def get_items(self):
-        messages = []
-        with self.lock:  # pylint: disable=not-context-manager
-            if self.messages:
-                messages = self.messages
-                self.messages = []
-        return messages
-
-    def add_item(self, *, item, cursor):
-        with self.lock:  # pylint: disable=not-context-manager
-            self.messages.append((item, cursor))
-            self.last_journal_msg_time = time.monotonic()
-            self.cursor = cursor
-
-        self.entry_num += 1
-        self.total_size += len(item)
 
 
 class JournalReader(Tagged):
